% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot_105.R
\docType{data}
\name{annot_105}
\alias{annot_105}
\title{Annotations file}
\format{
A 55416 x 5 data frame with gene annotations from Ensembl v105.
\describe{
\item{id}{Ensembl gene IDs}
\item{symbol}{Associated MGI symbol}
\item{chr}{Chromosome of the gene}
\item{start}{Start position (bp) on chromosome}
\item{end}{End position (bp) on chromosome}
}
}
\source{
ensembl.org
}
\usage{
data(annot_105)
}
\description{
Ensembl v105 annotations.
}
