% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Interpolate peak positions and attach annotations}
\usage{
annotatePeaks(map, peaks, biomart, localRange = 1e+07)
}
\arguments{
\item{map}{map_list from \link{mapQTL}.}

\item{peaks}{Un-annotated peaks.}

\item{biomart}{Annotations file. Path to location or object.}

\item{localRange}{is defined as "local". Default is 10e6.}
}
\value{
List of annotated peaks for each tissue.
}
\description{
Interpolate peak positions and attach annotations
}
