% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_meta.R
\docType{data}
\name{demo_meta}
\alias{demo_meta}
\title{Demo Metadata}
\format{
A 243 x 2 data frame with metadata information for each sample.
\describe{
\item{ID}{Sample IDs}
\item{sex}{sex of the sample origin}
}
}
\source{
Skelly et al., Mapping the Effects of Genetic Variation on Chromatin State and Gene Expression Reveals Loci That Control Ground State Pluripotency, Cell Stem Cell (2020), https://doi.org/10.1016/j.stem.2020.07.005
}
\usage{
data(demo_annot)
}
\description{
Sample metadata to be used for vignette.
}
