% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo_probs.R
\docType{data}
\name{demo_probs}
\alias{demo_probs}
\title{Demo Probabilities}
\format{
A list of genoprobs for each sample/marker separated by chromosome
\describe{
\item{Outer List}{Chromosome}
\item{Inner List}{3D dataset with samples x haplotype x marker}
}
}
\source{
Skelly et al., Mapping the Effects of Genetic Variation on Chromatin
State and Gene Expression Reveals Loci That Control Ground State
Pluripotency, Cell Stem Cell (2020),
https://doi.org/10.1016/j.stem.2020.07.005
}
\usage{
data(demo_probs)
}
\description{
Sample genotype probabilities to be used for vignette.
}
