% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile.R
\docType{data}
\name{gridfile}
\alias{gridfile}
\title{Grid Map}
\format{
A 74165 x 4 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
\item{bp}{Marker position in basepairs}
}
}
\source{
Updated GBRS software, see "https://github.com/TheJacksonLaboratory/cs-nf-pipelines/wiki/GBRS-Pipeline-ReadMe" for details.
}
\usage{
data(gridfile)
}
\description{
75k grid map for mm39 genome assembly.
}
