% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile250.R
\docType{data}
\name{gridfile250}
\alias{gridfile250}
\title{Grid Map}
\format{
A 250000 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
See https://github.com/TheJacksonLaboratory/jds-nf-workflows/wiki/Low-Coverage-WGS-HR-ReadMe for more information about the pipeline.
}
\usage{
data(gridfile250)
}
\description{
Genetic and Physical Maps to be used with genotypes from the JAX lcwgs-hr nextflow pipeline.
}
