% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile69k.R
\docType{data}
\name{gridfile69k}
\alias{gridfile69k}
\title{Grid Map}
\format{
A 69005 x 4 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
Updated GBRS software, see "https://github.com/TheJacksonLaboratory/cs-nf-pipelines/wiki/GBRS-Pipeline-ReadMe" for details.
}
\usage{
data(gridfile69k)
}
\description{
69k genetic and physical map for mm39 genome assembly.
}
