% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile_GM.R
\docType{data}
\name{gridfile_GM}
\alias{gridfile_GM}
\title{Grid Map}
\format{
A 143259 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
https://raw.githubusercontent.com/kbroman/MUGAarrays/main/UWisc/gm_uwisc_v4.csv.
}
\usage{
data(gridfile_GM)
}
\description{
GigaMUGA genetic and physical maps for mm39 genome assembly.
}
