% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile_mini.R
\docType{data}
\name{gridfile_mini}
\alias{gridfile_mini}
\title{Grid Map}
\format{
A 11125 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
https://raw.githubusercontent.com/kbroman/MUGAarrays/refs/heads/main/UWisc/mini_uwisc_v5.csv.
}
\usage{
data(gridfile_mini)
}
\description{
miniMUGA genetic and physical map for mm39 genome assembly.
}
