% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile_MM.R
\docType{data}
\name{gridfile_MM}
\alias{gridfile_MM}
\title{Grid Map}
\format{
A 77808 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
https://raw.githubusercontent.com/kbroman/MUGAarrays/refs/heads/main/UWisc/mm_uwisc_v4.csv.
}
\usage{
data(gridfile_MM)
}
\description{
MegaMUGA genetic and physical map for mm39 genome assembly.
}
