% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile_quilt.R
\docType{data}
\name{gridfile_quilt}
\alias{gridfile_quilt}
\title{Grid Map}
\format{
A 999999 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
See https://www.biorxiv.org/content/10.1101/2025.05.19.654056v1 for more information on Genotyping by lcWGS.
}
\usage{
data(gridfile_quilt)
}
\description{
Genetic and Physical Maps to be used with genotypes from Genotyping by lcWGS.
}
