% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_modiFinder.R
\name{gxe_modiFinder}
\alias{gxe_modiFinder}
\title{Prepare and run mediation for delta G x E peaks}
\usage{
gxe_modiFinder(
  peaks,
  mapping,
  by = "delta",
  sigLOD = 7.5,
  annots,
  outdir = NULL,
  med_out = "gxe_mediation.rds",
  total_cores = NULL,
  save = "sr",
  distOnly = TRUE,
  hsOnly = FALSE,
  BPPARAM = BiocParallel::SerialParam()
)
}
\arguments{
\item{peaks}{List of dataframes containing delta G x E peaks (name 'delta')}

\item{mapping}{Mapping list from \code{gxeQTL}. Needs to contain mapping information
for at minimum the delta condition, and the condition being used for mediation}

\item{by}{String indicating the name of the condition to be used for mediation.
Default 'delta'.}

\item{sigLOD}{Significant LOD threshold to use to filter phenotypes for
mediation. Default is 7.5}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{outdir}{Directory to save output files. Default is \code{NULL}.}

\item{med_out}{String indicating the name of the output file containing
mediation results for mediation within a phenotype. This file will be
saved in \code{.rds} format and used for downstream analysis and visualization.
Should end in \code{.rds}. Default is "\code{delta_mediation.rds}"}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}

\item{distOnly}{Logical. Mediate only the distal peaks? Default is \code{TRUE}.}

\item{hsOnly}{Logical. Mediate only on peaks identified within a hotspot.
Default is \code{FALSE}.}

\item{BPPARAM}{BiocParallel Parameter}
}
\value{
A list containing mediation results for the peaks called from the delta G x E method.
Mediated by 'delta', '\if{html}{\out{<ctrl>}}', or '\if{html}{\out{<env>}}'.
}
\description{
Prepare and run mediation for delta G x E peaks
}
