% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsHapEffects.R
\name{hsHapEffects}
\alias{hsHapEffects}
\title{Map founder haplotype effects for a given hotspot.}
\usage{
hsHapEffects(
  effects,
  tbands,
  chromosome,
  tissue,
  sigLOD,
  hsNum = 1,
  pop = "do",
  founders = NULL,
  palette = NULL,
  topFeats = NULL,
  psave = TRUE,
  pname = NULL,
  outdir = NULL,
  vert = FALSE,
  ...
)
}
\arguments{
\item{effects}{Effects object \code{qtl_effects}.}

\item{tbands}{List of transband locations (\code{bands.rna} list from
\code{transbands} function)}

\item{chromosome}{Chromosome that the transband (hotspot) is present on.}

\item{tissue}{Tissue that the transband (hotspot) is present in}

\item{sigLOD}{Significant LOD threshold to use for filtering phenotypes.}

\item{hsNum}{If there are multiple hotspots on a chromosome,
indicate which one. Default is 1.}

\item{pop}{One of \code{c("do", "cc", "other")} to indicate founder population.
Default is "\code{do}".}

\item{founders}{If \code{pop == "other"}, list of founders in haplotype order.}

\item{palette}{Founder color map.}

\item{topFeats}{Optional. Top hotspot features based on PCA analysis.}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{pname}{File name to save plot (needs to end in \code{.png}). Default is
\verb{haplotype_effects_<tissue>_transband_<hsNum>_chromosome_<chromosome>.png}}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{vert}{Logical. Rotate plot to vertical orientation. Default \code{FALSE}}

\item{...}{Additional arguments to pass to ComplexHeatmap}
}
\description{
Map founder haplotype effects for a given hotspot.
}
