% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcPeakPlot.R
\name{hsPeakPlot}
\alias{hsPeakPlot}
\title{Plot an individual peak or Principal Component of a hotspot.}
\usage{
hsPeakPlot(
  mapping,
  feats,
  tbands,
  chromosome,
  tissue,
  candidateMed = NULL,
  hsNum = 1,
  color = "#0073C2FF",
  pc = FALSE,
  wag = FALSE,
  sigLOD,
  topPC = 0.5,
  psave = TRUE,
  pname = NULL,
  outdir = NULL,
  ...
)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}.}

\item{feats}{List of phenotypes to include.
Individual/Candidate Mediator Only: Local to hotspot. PC: Targets of hotspot.}

\item{tbands}{List of transband locations (\code{bands.rna} list from
\code{transbands} function).}

\item{chromosome}{Chromosome that the transband (hotspot) is present on.}

\item{tissue}{Tissue that the transband (hotspot) is present in.}

\item{candidateMed}{Dataframe of candidate mediator information
(can be from annotations). Columns must include "id", "symbol",
"start", "end".}

\item{hsNum}{If there are multiple hotspots on a chromosome,
indicate which one. Default is 1.}

\item{color}{Plot color. Default is "#0073C2FF".}

\item{pc}{Logical. Plot the first principal component of the targets.
Default \code{FALSE}.}

\item{wag}{Logical. Include \code{tailWag} effects plot in final.
Default is \code{FALSE}.}

\item{sigLOD}{Significant LOD threshold to use for filtering phenotypes.}

\item{topPC}{Proportion of top hotspot contributers to be identified with
Principal Component Analysis. Default is 0.5.}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{...}{Additional arguments}
}
\value{
ggplot/patchwork object: Peak plot (with or without phenotype
effects). If plotting the Principal Component of the hotspot, also includes
a list of the top proportion of phenotypes contributing to the hotspot.
}
\description{
Plot an individual peak or Principal Component of a hotspot.
}
