% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_bp.R
\name{interp_bp}
\alias{interp_bp}
\title{interp_bp}
\usage{
interp_bp(df, genmap, physmap)
}
\arguments{
\item{df}{A peak dataframe with the peak position in cM.}

\item{genmap}{Genetic map - found in \code{maps_list} (gmap).}

\item{physmap}{Physical map - found in \code{maps_list} (pmap).}
}
\value{
Peak dataframe with interpolated bp column appended
}
\description{
Interpolate bp peak positions to bp based on marker maps
}
