% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_cM.R
\name{interp_cM}
\alias{interp_cM}
\title{interp_cM}
\usage{
interp_cM(df, genmap, physmap)
}
\arguments{
\item{df}{A peak dataframe with the peak position in bp.}

\item{genmap}{Genetic map - found in \code{maps_list} (gmap).}

\item{physmap}{Physical map - found in \code{maps_list} (pmap).}
}
\value{
Peak dataframe with interpolated bp column appended
}
\description{
Interpolate bp peak positions to cM based on marker maps.
}
