% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modiFinder.R
\name{modiFinder}
\alias{modiFinder}
\title{Prepare and run mediation for a set of QTL peaks}
\usage{
modiFinder(
  peaks,
  mapping,
  sigLOD = 7.5,
  annots,
  outdir = NULL,
  med_out = "mediation.rds",
  total_cores = NULL,
  save = "sr",
  distOnly = TRUE,
  hsOnly = FALSE
)
}
\arguments{
\item{peaks}{List of dataframes containing QTL peaks for each tissue, or
full path to \code{.rds} containing one.}

\item{mapping}{Mapping list from \code{mapQTL}, or full path to \code{.rds}
containing one.}

\item{sigLOD}{Significant LOD threshold to use to filter phenotypes for
mediation. Default is 7.5}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{outdir}{Directory to save output files. Default is \code{NULL}.}

\item{med_out}{String indicating the name of the output file containing
mediation results for mediation within a phenotype. This file will be
saved in \code{.rds} format and used for downstream analysis and visualization.
Should end in \code{.rds}. Default is "\code{mediation.rds}"}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}

\item{distOnly}{Logical. Mediate only the distal peaks? Default is \code{TRUE}.}

\item{hsOnly}{Logical. Mediate only on peaks identified within a hotspot.
Default is \code{FALSE}.}
}
\value{
A list containing mediation results for each tissue
}
\description{
Prepare and run mediation for a set of QTL peaks
}
