% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mugaprobs.R
\name{mugaprobs}
\alias{mugaprobs}
\title{Convert and process MUGA probabilities to qtl2 format}
\usage{
mugaprobs(
  type = "GM",
  covarLoc,
  covar_file,
  i.files,
  genoPrefix = "gm4qtl2",
  probsOut = "muga_interpolated_genoprobs.rds",
  saveDir = getwd(),
  tissues = c()
)
}
\arguments{
\item{type}{string indicating GigaMUGA (GM) or MegaMUGA (MM), default is "GM"}

\item{covarLoc}{location of covariate file}

\item{covar_file}{covariate file including at minimum the sex (sex) and generation (ngen) of each sample, this needs to be a .csv file}

\item{i.files}{either a string of the directory where the chromosome specific genotype files are or a list of final report files to process - if passing the final report files they need to be either unzipped or in .gz format}

\item{genoPrefix}{prefix for the chromosome specific genotype files (excluding "_geno")}

\item{probsOut}{file name to save probabilities, default is "muga_interpolated_genoprobs.rds".}

\item{saveDir}{complete path (\strong{not} relative) to save directory. Default is current directory.}

\item{tissues}{list of tissues included in analysis. If left blank tissue will be set to "a".}
}
\value{
none
}
\description{
Converts MUGA genotyping final reports into a 3D array format compatible with \code{qtl2}, organized by tissue and sample, optionally saves the result as an RDS file
}
