% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_modiFinder.R
\name{multi_modiFinder}
\alias{multi_modiFinder}
\title{Prepare and run between phenotype mediation for a set of QTL peaks with
provided expression data.}
\usage{
multi_modiFinder(
  peaks,
  mapping,
  exprZ,
  suggLOD = 7,
  annots,
  outdir = NULL,
  med_out = "multi_pheno_mediation.rds",
  total_cores = NULL,
  save = "sr"
)
}
\arguments{
\item{peaks}{List of dataframes containing QTL peaks for each tissue, or
full path pointing to saved peaks object.}

\item{mapping}{Mapping list from \code{mapQTL}, or full path to \code{.rds}
containing one.}

\item{exprZ}{rankZ transformed expression to use for the mediation.}

\item{suggLOD}{Significant LOD threshold to use to filter phenotypes for
mediation. Default is 7.5}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{outdir}{Directory to save output files. Default is \code{NULL}.}

\item{med_out}{String indicating the name of the output file containing
mediation results for mediation between phenotypes. This file will be saved
in \code{.rds} format and used for downstream analysis and visualization. Should
end in \code{.rds}. Default is "\code{multi_pheno_mediation.rds}"}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}
}
\value{
A list containing mediation results for each tissue
}
\description{
Mediation scan between two different phenotypes. Provide peaks and mapping
for the phenotype to mediate, and the rank Z-transformed phenotype
quantification to use as mediators.
}
