% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{peak_plot}
\alias{peak_plot}
\title{Plot peaks associated with specific genes.}
\usage{
peak_plot(
  mapping,
  tissue,
  pheno,
  saveDir,
  pop = "do",
  annots = NULL,
  effects = FALSE,
  founders = NULL,
  palette = NULL
)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}}

\item{tissue}{Tissue to derive plots from}

\item{pheno}{Phenotype to derive plots from}

\item{saveDir}{Directory to save plots to}

\item{pop}{Are you using a "do", "cc", or "other" population? Default is "do"}

\item{annots}{Annotations (example \code{annot_105}), required if plotting effects on chromosome}

\item{effects}{Boolean - Plotting effects on chromosome, default FALSE}

\item{founders}{If not using DO or CC mice, what are the founders of your population?}

\item{palette}{If using a population with more than 8 founders please provide colors to go with each founder.}
}
\value{
None
}
\description{
Plot peaks associated with specific genes.
}
