% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_eqtlmap}
\alias{plot_eqtlmap}
\title{Plot QTL maps (peak vs gene)}
\usage{
plot_eqtlmap(
  map_dat,
  peaks,
  sigLOD = 7.5,
  outdir = NULL,
  pname = NULL,
  psave = TRUE,
  unit = "bp",
  map_col = "#0073C2FF"
)
}
\arguments{
\item{map_dat}{\code{map_dat2} from \code{mapQTL} mapping list.}

\item{peaks}{List of dataframes containing QTL peaks for each tissue
(annotated).}

\item{sigLOD}{Significant LOD threshold to use for filtering phenotypes.
Default is 7.5}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).
Default is "eqtl_map_LOD\if{html}{\out{<sigLOD>}}_\if{html}{\out{<tissue>}}.png".}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{unit}{One of \code{c("bp", "mbp")}; annotation position units.
Default is "bp".}

\item{map_col}{Plot color. Default is "#0073C2FF"}
}
\value{
A list of peak maps (ggplot objects) for each tissue
}
\description{
Plot QTL maps (peak vs gene)
}
