% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_effects.R
\name{qtl_effects}
\alias{qtl_effects}
\title{Calculate founder effects for significant peaks}
\usage{
qtl_effects(
  peaks,
  mapping,
  suggLOD = 6,
  outdir = NULL,
  effects_out = "effects.rds",
  total_cores = NULL,
  save = "sr"
)
}
\arguments{
\item{peaks}{List of dataframes containing annotated QTL peaks for each
tissue, or full path to \code{.rds} containing one.}

\item{mapping}{Mapping list from \code{mapQTL}, or full path to \code{.rds}
containing one.}

\item{suggLOD}{Suggestive LOD threshold to use for filtering phenotypes.
Default is 6.}

\item{outdir}{Directory to save effects output files. Default is \code{NULL}.}

\item{effects_out}{String indicating the name of the output file containing
founder haplotype effects results. This file will be saved in \code{.rds} format
and used for downstream analysis and visualization. Should end in \code{.rds}.
Default is "\code{effects.rds}"}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}
}
\value{
A list containing: \itemize{
\item{effects_blup}{QTL effect BLUPs from scan along one chromosome. Output
from \code{\link[qtl2:scan1blup]{qtl2::scan1blup()}}}
\item{peaks}{annotated peaks with LOD scores above suggestive threshold}}
}
\description{
Calculate founder effects for significant peaks
}
