% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QTLretrievR-package.R
\docType{package}
\name{QTLretrievR-package}
\alias{QTLretrievR}
\alias{QTLretrievR-package}
\title{QTLretrievR: Calculate the Map, Peaks, Mediation, and Effects for Quantitative Trait Loci}
\description{
An integration of 'qtl2' and 'intermediate' to map genotypes to probes (genoprobs), calculate Quantitative Trait Loci (QTL) peaks, identify mediators of the peaks, and identify the phenotypic effects within experimental data from diversity outbred (DO) mice.
}
\author{
\strong{Maintainer}: Hannah B Dewey \email{hannah.dewey@jax.org} (\href{https://orcid.org/0000-0002-3009-0452}{ORCID})

Authors:
\itemize{
  \item Selcan Aydin \email{selcan.aydin@jax.org} (\href{https://orcid.org/0000-0002-0457-9641}{ORCID})
  \item Dan Skelly \email{dan.skelly@jax.org} (\href{https://orcid.org/0000-0002-2329-2216}{ORCID})
}

}
\keyword{internal}
