% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_mediate.R
\name{run_mediate}
\alias{run_mediate}
\title{Prepare and run mediation for a set of QTL peaks}
\usage{
run_mediate(
  peaks,
  mapping,
  suggLOD = 7,
  outdir,
  biomart,
  med_out,
  total_cores = NULL,
  save = "sr"
)
}
\arguments{
\item{peaks}{List of dataframes containing QTL peaks for each tissue}

\item{mapping}{List of relevant mapping information for overall project}

\item{suggLOD}{Suggestive LOD to use as filter for mediation. Default is 7.}

\item{outdir}{String of path to output directory where mediation lists will be saved.}

\item{biomart}{String pointing to annotations file or annotations object.}

\item{med_out}{Output file name to save mediation results for later use}

\item{total_cores}{Number of available cores to use for parallelization. Default is NULL.}

\item{save}{Should files be saved, returned, or both. Default is "sr" (save and return). To save only use "so", to return only use "ro".}
}
\value{
A list containing mediation results for each tissue
}
\description{
Prepare and run mediation for a set of QTL peaks
}
