% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_counts.R
\docType{data}
\name{sample_counts}
\alias{sample_counts}
\title{Sample Counts}
\format{
A 1000 x 184 matrix of normalized gene expression with each column as a different sample and genes in the rows.
\describe{
\item{rows}{Gene names as Ensembl IDs}
\item{cols}{Individual sample IDs}
}
}
\source{
Skelly et al., Mapping the Effects of Genetic Variation on Chromatin State and Gene Expression Reveals Loci That Control Ground State Pluripotency, Cell Stem Cell (2020), https://doi.org/10.1016/j.stem.2020.07.005
}
\usage{
data(sample_counts)
}
\description{
Downsampled mESC gene expression counts to be used for vignette.
}
