% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{split_map}
\alias{split_map}
\title{Background functions for QTLretrievR that will be called by the primary functions}
\usage{
split_map(map, chr_names = NULL)
}
\description{
Background functions for QTLretrievR that will be called by the primary functions
}
