% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tailWag.R
\name{tailWag}
\alias{tailWag}
\title{Create a plot with founder effects for a given phenotype}
\usage{
tailWag(
  effects,
  tissue,
  feat,
  chromosome,
  color = "#0073C2FF",
  psave = TRUE,
  pname = NULL,
  outdir = NULL,
  pop = "do",
  founders = NULL,
  symbol = TRUE
)
}
\arguments{
\item{effects}{Effects object (output from \code{qtl_effects} function).}

\item{tissue}{String indicating which tissue to plot the founder effects for.}

\item{feat}{Individual phenotype to plot effects of.}

\item{chromosome}{Chromosome that the peak is present on.}

\item{color}{Plot color. Default is "#0073C2FF".}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).
Default is \verb{wag_effects_<tissue>_<feat>.png}}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{pop}{One of \code{c("do", "cc", "other")} to indicate founder population.
Default is "\code{do}".}

\item{founders}{If \code{pop == "other"}, list of founders in haplotype order.}

\item{symbol}{Logical. The phenotype to be plotted is passed as a symbol
instead of phenotype id. Default \code{TRUE}.}
}
\value{
ggplot object of the effects plot
}
\description{
Create a plot with founder effects for a given phenotype
}
