% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transbands.R
\name{transbands}
\alias{transbands}
\title{Identify distal hotspots above a suggestive and significant LOD score.}
\usage{
transbands(
  map_dat,
  peaks,
  sigLOD = 7.5,
  suggLOD = 6,
  psave = TRUE,
  pname = NULL,
  outdir = NULL,
  color = "#0073C2FF"
)
}
\arguments{
\item{map_dat}{\code{map_dat2} from \code{mapQTL} mapping list.}

\item{peaks}{List of dataframes containing QTL peaks for each
tissue (annotated).}

\item{sigLOD}{Significant LOD threshold to use for filtering phenotypes.
Default is 7.5}

\item{suggLOD}{Suggestive LOD threshold to use for filtering phenotypes.
Default is 6.}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).
Default is \verb{hotspots_<tissue>.png}}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{color}{Plot color. Default is "#0073C2FF".}
}
\value{
A list containing:
\item{bands.rna}{
A list of tibbles containing hotspot information for each tissue.
Each tibble includes:
\describe{
\item{Chromosome}{Chromosome of the hotspot}
\item{Start/Stop}{Start and stop positions (bp) of the hotspot}
\item{Significant peaks}{Number of significant peaks in the hotspot}
\item{Suggestive peaks}{Number of suggestive peaks in the hotspot}
}
}
\item{trans_band_plots}{
A list of \code{ggplot2} objects (one per tissue) showing the number
of significant distal hotspot peaks.
}
}
\description{
Identify distal hotspots above a suggestive and significant LOD score.
}
