% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotatePeaks.R
\name{annotatePeaks}
\alias{annotatePeaks}
\title{Interpolate peak positions and attach annotations}
\usage{
annotatePeaks(mapping, peaks, annots, localRange = 1e+07)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}..}

\item{peaks}{List of un-annotated peaks}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{localRange}{Definition of "local" in bp. Default is 10e6..}
}
\value{
List of annotated peaks for each tissue.
}
\description{
Interpolate peak positions and attach annotations
}
