% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genoprobably.R
\name{genoprobably}
\alias{genoprobably}
\title{Convert GBRS tsv genome probabilities to genoprobs format.}
\usage{
genoprobably(
  outfile = "./gbrs_interpolated_genoprobs.rds",
  gbrsFileLoc,
  tissues = c(),
  gridfile = gridfile
)
}
\arguments{
\item{outfile}{Name of file to save genoprobs to. Defaults to "gbrs_interpolated_genoprobs.rds"}

\item{gbrsFileLoc}{File path to where the GBRS files are located.}

\item{tissues}{List of tissues included in analysis. If left blank tissue will be set to "a".}

\item{gridfile}{File location for genome grid. Defaults to object loaded with package for 75k grid.}
}
\value{
List of three dimensional genome probabilities
}
\description{
Convert GBRS tsv genome probabilities to genoprobs format.
}
