% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile.R, R/gridfile_quilt.R
\docType{data}
\name{gridfile}
\alias{gridfile}
\title{Grid Map}
\format{
A 74165 x 4 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}

A 999999 x 3 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
}
}
\source{
Updated GBRS software, see "https://github.com/TheJacksonLaboratory/cs-nf-pipelines/wiki/GBRS-Pipeline-ReadMe" for details.

See https://www.biorxiv.org/content/10.1101/2025.05.19.654056v1 for more information on Genotyping by lcWGS.
}
\usage{
data(gridfile)

data(gridfile_quilt)
}
\description{
75k genetic and physical map for mm39 genome assembly.

Genetic and Physical Maps to be used with genotypes from Genotyping by lcWGS.
}
