% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gridfile_MM.R
\docType{data}
\name{gridfile_MM}
\alias{gridfile_MM}
\title{Grid Map}
\format{
A 11125 x 4 data frame with marker names included as row names:
\describe{
\item{chr}{Chromosome name}
\item{pos}{Marker position in basepairs}
\item{cM}{Marker position in centiMorgan}
\item{bp}{Marker position in basepairs}
}
}
\source{
https://raw.githubusercontent.com/kbroman/MUGAarrays/refs/heads/main/UWisc/mini_uwisc_v5.csv.
}
\usage{
data(gridfile_MM)
}
\description{
MegaMUGA grid map for mm39 genome assembly.
}
