% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gxeQTL.R
\name{gxeQTL}
\alias{gxeQTL}
\title{Generate mapping data and peaks for G x E QTL analysis}
\usage{
gxeQTL(
  genoprobs,
  samp_meta,
  expr_mats,
  covar_factors,
  thrA = 5,
  thrX = 5,
  gridFile = gridfile,
  localRange = 2e+06,
  outdir = NULL,
  peaks_out = "gxe_peaks.rds",
  map_out = "gxe_map.rds",
  annots = NULL,
  total_cores = NULL,
  save = "sr",
  delta = FALSE,
  ctrl,
  env,
  rz = FALSE
)
}
\arguments{
\item{genoprobs}{Genotype probabilities object (qtl2-formatted), or path to
.rds file containing one.}

\item{samp_meta}{Sample metadata. Either a string pointing to the file, or
the object itself.}

\item{expr_mats}{List of normalized count matrices (objects), or character
paths to the file. One matrix per tissue. The order \emph{must match} the
tissue order in \code{genoprobs}.}

\item{covar_factors}{Additive covariate factors. These need to be columns
in the factor metadata.}

\item{thrA}{Minimum reported LOD threshold for autosomes. Default is 5.}

\item{thrX}{Minimum reported LOD threshold for X chromosome. Default is 5.}

\item{gridFile}{Genome Grid. Path to location or object. Defaults to
75k grid loaded with package.}

\item{localRange}{Definition of "local" in bp. Default is 2e6.}

\item{outdir}{Directory to save output files. Default is \code{NULL}.}

\item{peaks_out}{String indicating the name for output peaks file. This file
will be saved in \code{.rds} format and and be used as an input for downstream
analysis. Should end  in \code{.rds}. Default is "\code{gxe_peaks.rds}"}

\item{map_out}{String indicating the name of the output file containing
G x E QTL mapping results. This file will be saved in \code{.rds} format and
will be used in downstream analyses. Should end in \code{.rds}.
Default is "\code{gxe_map.rds}"}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}

\item{delta}{Logical. The delta method for G x E analysis (env - ctrl)
should be used. Default is \code{FALSE}.}

\item{ctrl}{String indicating your control or background gene name
(ex: "ctrl" or "CTRL").}

\item{env}{String indicating your exposed/treated samples
(ex: "trt" or "treated" or "<your_treatment_here>").}

\item{rz}{Logical. Set to \code{TRUE} if expression data is already
rankZ-transformed. Default is \code{FALSE}.}
}
\value{
A list containing: \itemize{
\item{maps_list}{A list of dataframes and lists to that can be used for
future analyses and in other functions \itemize{
\item{qtlprobs}{Genome probabilities in qtl2format}
\item{covar_list}{list of covariate matrices for each tissue}
\item{expr_list}{Original normalized expression data for each tissue}
\item{exprZ_list}{Rank Z normalized expression data for each tissue}
\item{kinship_loco}{Kisnhip Matrix calculated using the "loco" option
in \code{qtl2::calc_kinship}}
\item{gmap}{Genomic map of markers}
\item{map_dat2}{Combined genomic and physical map of markers}
\item{pmap}{Physical map of markers}
\item{tissue_samp}{Metadata broken down for each tissue}}}
\item{peaks_list}{A list of peaks list for each tissue.}}
}
\description{
This function performs G x E QTL mapping across multiple tissues,
generating genome-wide LOD scores and identifying significant peaks.
It supports parallel processing and flexible input formats for expression
data and covariates.
}
