% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOD_thld.R
\name{LOD_thld}
\alias{LOD_thld}
\title{Determine significant and suggestive LOD thresholds}
\usage{
LOD_thld(
  mapping,
  tissue,
  annots = NULL,
  n.gene = 10,
  n.perm = 1000,
  batch.size = 5
)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}, or full path to \code{.rds}
containing one.}

\item{tissue}{Tissue to determine thresholds for.}

\item{annots}{Data frame of phenotype annotations for filtering to
autosomal phenotypes. Columns must include "id", "symbol", "start", "end".}

\item{n.gene}{Number of phenotypes to run permutations on. Default is 10.}

\item{n.perm}{Number of permutations to run per phenotype.  Default is 1000.}

\item{batch.size}{Number of genes in each parallelized batch.}
}
\value{
A list containing:
\describe{
\item{perms}{A \code{scan1perm} object of permutations for all selected genes.}
\item{thresholds}{A list with two elements:
\describe{
\item{significant}{Median LOD threshold at alpha = 0.05.}
\item{suggestive}{Median LOD threshold at alpha = 0.4.}
}
}
}
}
\description{
Determine significant and suggestive LOD thresholds for your QTLs, using
\code{scan1perm} from \code{qtl2}. Select the number of genes to permute, the number of
permutations, and the number of genes to run in each batch.
}
