% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOD_thld.R
\name{LOD_thld}
\alias{LOD_thld}
\title{Determine significant and suggestive LOD thresholds}
\usage{
LOD_thld(mapping, tissue, annots, n.gene = 10, n.perm = 1000, batch.size = 5)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}}

\item{tissue}{Tissue to determine thresholds for}

\item{annots}{annotations}

\item{n.gene}{Number of genes to permute - default 10}

\item{n.perm}{Number of permutations to run - default 1000}

\item{batch.size}{Number of genes in each parallelized batch}
}
\value{
A list containing: \itemize {
\item{\code{scan1perm} object of permutations for all selected genes}
\item{A list of: \itemize {
\item{Significant LOD threshold (median of alpha = 0.05 summary for all genes)}
\item{Suggestive LOD threshold (median of alpha = 0.4 summary for all genes)}}}}
}
\description{
Determine significant and suggestive LOD thresholds for your QTLs, using \code{scan1perm} from \code{qtl2}. Select the number of genes to permute, the number of permutations, and the number of genes to run in each batch.
}
