% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_mediate.R
\name{modiFinder}
\alias{modiFinder}
\title{Prepare and run mediation for a set of QTL peaks}
\usage{
modiFinder(
  peaks,
  mapping,
  suggLOD = 7,
  annots,
  outdir = NULL,
  med_out = "mediation.rds",
  total_cores = NULL,
  save = "sr",
  distOnly = T
)
}
\arguments{
\item{peaks}{List of dataframes containing QTL peaks for each tissue, or full path pointing to saved peaks object.}

\item{mapping}{List of relevant mapping information for overall project, or full path pointing to saved peaks object.}

\item{suggLOD}{Suggestive LOD to use as filter for mediation. Default is 7.}

\item{annots}{String pointing to annotations file or annotations object.}

\item{outdir}{String of path to output directory where mediation lists will be saved. Default is NULL.}

\item{med_out}{String indicating the name for output mediation file. Should end in ".rds". Default is "mediation.rds"}

\item{total_cores}{Number of available cores to use for parallelization. Default is NULL.}

\item{save}{Should files be saved, returned, or both. Default is "sr" (save and return). To save only use "so", to return only use "ro".}

\item{distOnly}{Logical. Mediate only the distal peaks? Default is TRUE}
}
\value{
A list containing mediation results for each tissue
}
\description{
Prepare and run mediation for a set of QTL peaks
}
