% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_mediate.R
\name{multi_modiFinder}
\alias{multi_modiFinder}
\title{Prepare and run between phenotype mediation for a set of QTL peaks with provided expression data.}
\usage{
multi_modiFinder(
  peaks,
  mapping,
  exprZ,
  suggLOD = 7,
  annots,
  outdir = NULL,
  med_out = "multi_pheno_mediation.rds",
  total_cores = NULL,
  save = "sr"
)
}
\arguments{
\item{peaks}{List of dataframes containing QTL peaks for each tissue, or full path pointing to saved peaks object.}

\item{mapping}{List of relevant mapping information for overall project, or full path pointing to saved peaks object.}

\item{exprZ}{rankZ transformed expression to use for the mediation}

\item{suggLOD}{Suggestive LOD to use as filter for mediation. Default is 7.}

\item{annots}{String pointing to annotations file or annotations object (for mediators).}

\item{outdir}{String of path to output directory where mediation lists will be saved. Default is NULL.}

\item{med_out}{String indicating the name for output mediation file. Should end in ".rds". Default is "multi_pheno_mediation.rds"}

\item{total_cores}{Number of available cores to use for parallelization. Default is NULL.}

\item{save}{Should files be saved, returned, or both. Default is "sr" (save and return). To save only use "so", to return only use "ro".}
}
\value{
A list containing mediation results for each tissue
}
\description{
Mediation scan between two different phenotypes. Provide peaks and mapping for the phenotype to mediate, and the rank Z-transformed phenotype quantification to use as mediators.
}
