% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{peak_plot}
\alias{peak_plot}
\title{Plot peaks associated with specific genes.}
\usage{
peak_plot(
  mapping,
  tissue,
  pheno,
  pop = "do",
  outdir = NULL,
  pname = NULL,
  psave = TRUE,
  chromosome = NULL,
  effects = FALSE,
  founders = NULL,
  palette = NULL
)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}, or full path to \code{.rds} containing
one.}

\item{tissue}{Tissue to derive plots from}

\item{pheno}{Phenotype to plot}

\item{pop}{One of \code{c("do", "cc", "other")} to indicate founder population.
Default is "\code{do}".}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{chromosome}{Chromosome that the peaks is present on.}

\item{effects}{Logical. Plotting effects on chromosome. Default is \code{FALSE}.}

\item{founders}{If \code{pop == "other"}, list of founders in haplotype order.}

\item{palette}{Founder color map (> 8 founders).}
}
\value{
Plot of phenotype specific peaks, or phenotype specific peak with
effects
}
\description{
Plot peaks associated with specific genes.
}
