% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_plot.R
\name{peak_plot}
\alias{peak_plot}
\title{Plot peaks associated with specific genes.}
\usage{
peak_plot(
  mapping,
  tissue,
  pheno,
  saveDir,
  pop = "do",
  annots = NULL,
  effects = FALSE,
  founders = NULL,
  palette = NULL
)
}
\arguments{
\item{mapping}{Mapping list from \code{mapQTL}}

\item{tissue}{Tissue to derive plots from}

\item{pheno}{Phenotype to derive plots from}

\item{saveDir}{Directory to save plots to}

\item{annots}{Annotations (example \code{annot_105}), required if plotting effects on chromosome}

\item{effects}{Boolean - Plotting effects on chromosome, default FALSE}
}
\value{
None
}
\description{
Plot peaks associated with specific genes.
}
