% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_eqtlmap}
\alias{plot_eqtlmap}
\title{Plot QTL maps (peak vs gene)}
\usage{
plot_eqtlmap(
  map_dat,
  peaks,
  sigLOD = 7.5,
  outdir = NULL,
  pname = NULL,
  psave = T,
  unit = "bp",
  map_col = "blue3"
)
}
\arguments{
\item{map_dat}{Mapping information for each marker used to determine genoprobs}

\item{peaks}{List of annotated peaks for each tissue}

\item{sigLOD}{Significant LOD threshold. Default is 7.5.}

\item{outdir}{String to output directory where plots should be saved. Default NULL.}

\item{pname}{String for plot name to save, must end in \code{.png}. Default NULL.}

\item{psave}{Whether or not to save plots to png. Default is TRUE.}

\item{unit}{Units for start/stop/midpoint from annotations. One of "bp" or "mbp". Default is "bp"}

\item{map_col}{Map Color. Default is "blue3"}
}
\value{
A list of peak maps (ggplot objects) for each tissue
}
\description{
Plot QTL maps (peak vs gene)
}
