% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_map.R
\name{plot_eqtlmap}
\alias{plot_eqtlmap}
\title{Plot QTL maps (peak vs gene)}
\usage{
plot_eqtlmap(
  map_dat,
  peaks,
  sigLOD = 7.5,
  outdir,
  outbase,
  psave = T,
  unit = "bp"
)
}
\arguments{
\item{map_dat}{Mapping information for each marker used to determine genoprobs}

\item{peaks}{List of annotated peaks for each tissue}

\item{sigLOD}{Significant LOD threshold. Default is 7.5.}

\item{outdir}{String to output directory where plots should be saved}

\item{outbase}{File name to save plots to}

\item{psave}{Whether or not to save plots to png. Default is TRUE.}

\item{unit}{Units for start/stop/midpoint from annotations. One of "bp" or "mbp".
Default is "bp"}
}
\value{
A list of peak maps (ggplot objects) for each tissue
}
\description{
Plot QTL maps (peak vs gene)
}
