% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qtl_effects.R
\name{qtl_effects}
\alias{qtl_effects}
\title{Calculate founder effects for significant peaks}
\usage{
qtl_effects(mapping, peaks, suggLOD = 8, outdir, outfile, total_cores)
}
\arguments{
\item{mapping}{List of relevant mapping information for overall project}

\item{peaks}{List of dataframes containing QTL peaks for each tissue}

\item{suggLOD}{Suggestive LOD threshold. Default is 7.}

\item{outdir}{String of path to output directory where effects lists will be saved.}

\item{outfile}{Output file name to save mediation results for later use}

\item{total_cores}{Number of available cores to use. Default is NULL.}
}
\value{
A list containing: \itemize{
\item{effects_blup}{QTL effect BLUPs from scan along one chromosome. Output from \code{\link[qtl2:scan1blup]{qtl2::scan1blup()}}}
\item{effects_std}{QTL effect coefficients from scan along one chromosme. Output from \code{\link[qtl2:scan1coef]{qtl2::scan1coef()}}}
\item{peaks}{annotated peaks with LOD scores above suggestive threshold}}
}
\description{
Calculate founder effects for significant peaks
}
