% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runQTL.R
\name{runQTL}
\alias{runQTL}
\title{Wrapper function to generate mapping, peaks, mediation, and effects data}
\usage{
runQTL(
  geno_out = "gbrs_interpolated_genoprobs.rds",
  peaks_out = "peaks.rds",
  map_out = "mapping.rds",
  med_out = "mediation_res.rds",
  effects_out = "effects.rds",
  outdir = NULL,
  gbrs_fileLoc,
  metadata,
  expr_mats,
  covar_factors,
  annots,
  tissues = c(),
  gridFile = gridfile,
  total_cores = NULL,
  save_t = "sr",
  ...
)
}
\arguments{
\item{geno_out}{String indicating the name of the output file containing
interpolated genotype probabilities. This file will be saved in \code{.rds}
format and used for QTL mapping. Should end in \code{.rds}. Default is
"\code{gbrs_interpolated_genoprobs.rds}".}

\item{peaks_out}{String indicating the name for output peaks file.
This file will be saved in \code{.rds} format and will be used as an input
for downstream analysis. Should end  in \code{.rds}. Default is "\code{peaks.rds}"}

\item{map_out}{String indicating the name of the output file containing QTL
mapping results. This file will be saved in \code{.rds} format and will be used
in downstream analyses. Should end in \code{.rds}. Default is "\code{map.rds}"}

\item{med_out}{String indicating the name of the output file containing
mediation results for mediation within a phenotype. This file will be saved
in \code{.rds} format and used for downstream analysis and visualization. Should
end in \code{.rds}. Default is "\code{mediation.rds}"}

\item{effects_out}{String indicating the name of the output file containing
founder haplotype effects results. This file will be saved in \code{.rds}
format and used for downstream analysis and visualization. Should end in
\code{.rds}. Default is "\code{effects.rds}"}

\item{outdir}{Directory to save output files. Default is \code{NULL}.}

\item{gbrs_fileLoc}{Path to GBRS interpolated tsv files.}

\item{metadata}{Sample metadata. Either a string pointing to the file,
or the object itself.}

\item{expr_mats}{List of normalized count matrices (objects), or character
paths to the file. One matrix per tissue. The order \emph{must match} the
tissue order in \code{genoprobs}.}

\item{covar_factors}{Additive covariate factors. These need to be columns
in the sample metadata.}

\item{annots}{Annotations file. Contains mapping information for phenotypes.
Dataframe, or tsv. Columns must include "id", "symbol", "start", "end".}

\item{tissues}{Vector of strings indicating tissues or conditions in project.}

\item{gridFile}{Genome Grid. Path to location or object. Defaults to
75k grid loaded with package.}

\item{total_cores}{Number of available cores to use for parallelization.
Default is \code{NULL}.}

\item{save_t}{Indicates object return/save behavior. One of
\code{c("sr", "so", "ro")}; save & return, save only, return only.
Default is "sr".}

\item{...}{Additional arguments}
}
\value{
A list containing \itemize{
\item{peaks_list}{Unfiltered peaks for each tissue.}
\item{maps_list}{List of objects associated with mapping. See \link{mapQTL}
help for details.}
\item{res_list}{List containing mediation results for each tissue.}
\item{effects_res}{List of objects associated with effects. See \link{qtl_effects}
help for details.}}
}
\description{
Wrapper function to generate mapping, peaks, mediation, and effects data
}
