% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transbands.R
\name{transbands}
\alias{transbands}
\title{Identify distal hotspots above a suggestive and significant LOD score.}
\usage{
transbands(
  map_dat,
  peaks,
  sigLOD = 7.5,
  suggLOD = 6,
  psave = TRUE,
  pname = NULL,
  outdir = NULL,
  color = "#0073C2FF"
)
}
\arguments{
\item{map_dat}{\code{map_dat2} from \code{mapQTL} mapping list.}

\item{peaks}{List of dataframes containing QTL peaks for each
tissue (annotated).}

\item{sigLOD}{Significant LOD threshold to use for filtering phenotypes.
Default is 7.5}

\item{suggLOD}{Suggestive LOD threshold to use for filtering phenotypes.
Default is 6.}

\item{psave}{Logical. Save the plot as \code{.png}. Default \code{TRUE}.}

\item{pname}{File name to save plot (needs to end in \code{.png}).
Default is \verb{hotspots_<tissue>.png}}

\item{outdir}{Directory to save plots. Default is \code{NULL}.}

\item{color}{Plot color. Default is "#0073C2FF".}
}
\value{
A list containing: \itemize{
\item{bands.rna}{A list of tibbles containing hotspot information for
each tissue
Information includes the following for each hotspot:
\itemize{
\item Chromosome of the hotspot
\item Start and Stop (bp) of the hotspot
\item The number of significant peaks within the hotspot
\item The number of suggestive peaks within the hotspot }
\item{trans_band_plots}{A list of ggplot2 objects (one for each tissue),
showing the number of significant distal hotspot peaks.}}
}
}
\description{
Identify distal hotspots above a suggestive and significant LOD score.
}
