% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transbands.R
\name{transbands}
\alias{transbands}
\title{Identify distal hotspots above a suggestive and significant LOD score.}
\usage{
transbands(
  map_dat,
  peaks,
  sigLOD = 7.5,
  suggLOD = 6,
  outdir,
  psave = TRUE,
  color = "blue3"
)
}
\arguments{
\item{map_dat}{Mapping information for each marker used to determine genoprobs}

\item{peaks}{List of annotated peaks for each tissue}

\item{sigLOD}{Significant LOD threshold. Default is 7.5.}

\item{suggLOD}{Suggestive LOD threshold. Default is 6.}

\item{outdir}{String to output directory where plots should be saved}

\item{psave}{Whether or not to save plots to png. Default is TRUE.}

\item{color}{Map Color. Default is "blue3"}
}
\value{
A list containing: \itemize{
\item{bands.rna}{A list of tibbles containing hotspot information for each tissue
Information includes the following for each hotspot:
\itemize{
\item Chromosome of the hotspot
\item Start and Stop (bp) of the hotspot
\item The number of significant peaks within the hotspot
\item The number of suggestive peaks within the hotspot }
\item{trans_band_plots}{A list of ggplot2 objects (one for each tissue),
showing the number of significant distal hotspot peaks.}}
}
}
\description{
Identify distal hotspots above a suggestive and significant LOD score.
}
